requirements.txt
streamlit>=1.31
python-dotenv>=1.0
openai>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
guardrails-ai>=0.4.0
langsmith>=0.1.0